backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
psycopg2-binary==2.9.3
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0 